Move WeatherForecast to Application. Code clean-up
