+++ source/Clipify.Application/DependencyInjection.cs
@@ -1,6 +1,6 @@
-﻿using System.Reflection;
-using MediatR;
+﻿using MediatR;
 using Microsoft.Extensions.DependencyInjection;
+using System.Reflection;
 
 namespace Clipify.Application
 {
